requirements.txt
pillow-simd
docopt
